requirements.txt
# Python 3.13 uyumlu versiyonlar
librosa==0.10.1
numpy==1.26.0
scikit-learn==1.3.2
matplotlib==3.8.0
pandas==2.1.3
torch==2.1.0
torchaudio==2.1.0
PyYAML==6.0.1
jupyter==1.0.0
soundfile==0.12.1
noisereduce==2.0.1
scipy==1.11.3